Clean up the dataset endline1_biz and update example minimal
In the previous versions, we filter out the observations with
`any_biz_1` equals to 0, which means that we exclude the households
that has no business. However, further investigation shows that only
53  observations that don't have `biz_index_all_1`. We don't know yet
why the households with no business at all still have
`biz_index_all_1`. In order to include more data into the dataset, we
will only filter out the observations that don't even have
`biz_index_all_1` for now.

Exclude all the *area* and *index* variable in endline1_biz.

Also update the `split_ratio` in the minimal example so the same split
ratio would be used in both *rf* and *crf* methods.